feat: saving todo on db

+++ internal/app/todo.go
@@ -1,7 +1,7 @@
 package app
 
 type Todo struct {
-	Id          int
+	Id          int64
 	Name        string
 	Description string
 	Done        bool
@@ -25,7 +25,7 @@ func (todo *Todo) SetDescription(description string) *Todo {
 	return todo
 }
 
-func FindById(todos []Todo, id int) *Todo {
+func FindById(todos []Todo, id int64) *Todo {
 	for i := 0; i < len(todos); i++ {
 		if todos[i].Id == id {
 			return &todos[i]
